+++ CHANGES.md
@@ -1,6 +1,7 @@
 # Dev (2023-??-??) - ??
 
 ## Features/Changes
+* Compiler: global dead code elimination (Micah Cantor, #1503)
 * Compiler: change control-flow compilation strategy (#1496)
 * Compiler: Dead code elimination of unused references (#2076)
 * Compiler: reduce memory consumption (#1516)